Align the names of the test case files and move setkodata into the tests
