+++ test/upgrade/knativeeventing_postupgrade_test.go
@@ -1,4 +1,4 @@
-// +build preupgrade
+// +build postupgrade
 
 /*
 Copyright 2020 The Knative Authors
@@ -19,18 +19,19 @@ limitations under the License.
 package e2e
 
 import (
+	"os"
 	"testing"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
+	"knative.dev/operator/pkg/reconciler/common"
 	"knative.dev/operator/test"
 	"knative.dev/operator/test/client"
 	"knative.dev/operator/test/resources"
 	"knative.dev/pkg/test/logstream"
 )
 
-// TestKnativeEventingPreUpgrade verifies the KnativeEventing creation, before upgraded to the latest HEAD at master.
-func TestKnativeEventingPreUpgrade(t *testing.T) {
+// TestKnativeEventingUpgrade verifies the KnativeEventing creation, deployment recreation, and KnativeEventing deletion
+// after upgraded to the latest HEAD at master, with the latest generated manifest of KnativeEventing.
+func TestKnativeEventingUpgrade(t *testing.T) {
 	cancel := logstream.Start(t)
 	defer cancel()
 	clients := client.Setup(t)
@@ -42,18 +43,18 @@ func TestKnativeEventingPreUpgrade(t *testing.T) {
 
 	// Create a KnativeEventing
 	if _, err := resources.EnsureKnativeEventingExists(clients.KnativeEventing(), names); err != nil {
-		t.Fatalf("KnativeEventing %q failed to create: %v", names.KnativeEventing, err)
+		t.Fatalf("KnativeService %q failed to create: %v", names.KnativeEventing, err)
 	}
 
 	// Verify if resources match the requirement for the previous release before upgrade
 	t.Run("verify resources", func(t *testing.T) {
 		resources.AssertKEOperatorCRReadyStatus(t, clients, names)
-		keventing, err := clients.KnativeEventing().Get(names.KnativeEventing, metav1.GetOptions{})
-		if err != nil {
-			t.Fatalf("Failed to get KnativeEventing CR: %v", err)
-		}
-		// Based on the status.version, get the deployment resources.
-		expectedDeployments := resources.GetExpectedDeployments(t, keventing.Status.Version, "knative-eventing")
+		kcomponent := "knative-eventing"
+		resources.SetKodataDir()
+		defer os.Unsetenv(common.KoEnvKey)
+		version := common.GetLatestRelease(kcomponent)
+		// Based on the latest release version, get the deployment resources.
+		expectedDeployments := resources.GetExpectedDeployments(t, version, kcomponent)
 		resources.AssertKnativeDeploymentStatus(t, clients, names.Namespace, expectedDeployments)
 	})
 }
