+++ test/upgrade/knativeeventing_preupgrade_test.go
@@ -1,4 +1,4 @@
-// +build postupgrade
+// +build preupgrade
 
 /*
 Copyright 2020 The Knative Authors
@@ -19,6 +19,7 @@ limitations under the License.
 package e2e
 
 import (
+	"os"
 	"testing"
 
 	"knative.dev/operator/pkg/reconciler/common"
@@ -28,9 +29,8 @@ import (
 	"knative.dev/pkg/test/logstream"
 )
 
-// TestKnativeEventingUpgrade verifies the KnativeEventing creation, deployment recreation, and KnativeEventing deletion
-// after upgraded to the latest HEAD at master, with the latest generated manifest of KnativeEventing.
-func TestKnativeEventingUpgrade(t *testing.T) {
+// TestKnativeEventingPreUpgrade verifies the KnativeEventing creation, before upgraded to the latest HEAD at master.
+func TestKnativeEventingPreUpgrade(t *testing.T) {
 	cancel := logstream.Start(t)
 	defer cancel()
 	clients := client.Setup(t)
@@ -42,17 +42,20 @@ func TestKnativeEventingUpgrade(t *testing.T) {
 
 	// Create a KnativeEventing
 	if _, err := resources.EnsureKnativeEventingExists(clients.KnativeEventing(), names); err != nil {
-		t.Fatalf("KnativeService %q failed to create: %v", names.KnativeEventing, err)
+		t.Fatalf("KnativeEventing %q failed to create: %v", names.KnativeEventing, err)
 	}
 
 	// Verify if resources match the requirement for the previous release before upgrade
 	t.Run("verify resources", func(t *testing.T) {
 		resources.AssertKEOperatorCRReadyStatus(t, clients, names)
-		kcomponent := "knative-eventing"
+		keventing, err := clients.KnativeEventing().Get(names.KnativeEventing, metav1.GetOptions{})
+		if err != nil {
+			t.Fatalf("Failed to get KnativeEventing CR: %v", err)
+		}
 		resources.SetKodataDir()
-		version := common.GetLatestRelease(kcomponent)
-		// Based on the latest release version, get the deployment resources.
-		expectedDeployments := resources.GetExpectedDeployments(t, version, kcomponent)
+		defer os.Unsetenv(common.KoEnvKey)
+		// Based on the status.version, get the deployment resources.
+		expectedDeployments := resources.GetExpectedDeployments(t, keventing.Status.Version, "knative-eventing")
 		resources.AssertKnativeDeploymentStatus(t, clients, names.Namespace, expectedDeployments)
 	})
 }
